+++ R/ggemmeans.R
@@ -17,8 +17,22 @@ ggemmeans <- function(model,
   }
 
   # check arguments
-  type <- match.arg(type, choices = c("fe", "fe.zi", "re", "re.zi"))
+  type <- match.arg(type, choices = c("fe", "fixed", "count", "re", "random", "fe.zi", "zero_inflated", "re.zi", "zero_inflated_random"))
   model_name <- deparse(substitute(model))
+
+  type <- switch(
+    type,
+    "fixed" = ,
+    "count" = "fe",
+    "random" = "re",
+    "zi" = ,
+    "zero_inflated" = "fe.zi",
+    "zi_random" = ,
+    "zero_inflated_random" = "re.zi",
+    "survival" = "surv",
+    "cumulative_hazard" = "cumhaz"    ,
+    type
+  )
 
   # check if terms are a formula
   if (!missing(terms) && !is.null(terms) && inherits(terms, "formula")) {
@@ -91,7 +105,7 @@ ggemmeans <- function(model,
 
   attr(prediction_data, "continuous.group") <- attr(data_grid, "continuous.group")
 
-  if (model_info$is_ordinal | model_info$is_categorical) {
+  if (model_info$is_ordinal || model_info$is_categorical || model_info$is_multinomial) {
     colnames(prediction_data)[1] <- "response.level"
   }
 
@@ -149,7 +163,7 @@ ggemmeans <- function(model,
     "fixed-effects"
   else if (inherits(model, "gls"))
     "satterthwaite"
-  else if (model_info$is_ordinal | model_info$is_categorical)
+  else if (model_info$is_ordinal || model_info$is_categorical || model_info$is_multinomial)
     "prob"
   else if (model_info$is_zero_inflated && type %in% c("fe", "re") && inherits(model, "glmmTMB"))
     "link"