+++ NEWS.md
@@ -1,3 +1,16 @@
+# ggeffects 0.14.2
+
+## General
+
+* *ggeffects* now requires _glmmTMB_ version 1.0.0 or higher.
+* Added human-readable alias-options to the `type`-argument.
+
+## Bug fixes
+
+* Fixed issue when log-transformed predictors where held constant and their typical value was negative.
+* Fixed issue when plotting raw data to a plot with categorical predictor in the x-axis, which had numeric factor levels that did not start at `1`.
+* Fixed issues for model objects that used (log) transformed `offset()` terms.
+
 # ggeffects 0.14.1
 
 ## General